+++ src/list0/utils.py
@@ -2,4 +2,4 @@ import numpy as np
 
 
 def tour_cost(dist: np.ndarray, path: np.ndarray):
-    return np.sum(dist[path, np.roll(path, -1)], dtype=np.int64)
+    return np.sum(dist[path, np.roll(path, -1)], dtype=np.int32)
